Moved renderers to Content project. Abstracts Web project's dependency on Contentful.
