+++ src/Dfe.EarlyYearsQualification.Content/Extensions/ServiceCollectionExtensions.cs
@@ -1,0 +1,14 @@
+using Dfe.EarlyYearsQualification.Content.Renderers.Entities;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Dfe.EarlyYearsQualification.Content.Extensions;
+
+public static class ServiceCollectionExtensions
+{
+    public static IServiceCollection AddModelRenderers(this IServiceCollection services)
+    {
+        services.AddSingleton<IHtmlRenderer, HtmlModelRenderer>();
+
+        return services;
+    }
+}
