+++ src/Dfe.EarlyYearsQualification.Web/Program.cs
@@ -1,9 +1,9 @@
 using System.Diagnostics.CodeAnalysis;
 using Azure.Identity;
 using Contentful.AspNetCore;
+using Dfe.EarlyYearsQualification.Content.Extensions;
 using Dfe.EarlyYearsQualification.Content.Services;
 using Dfe.EarlyYearsQualification.Mock.Extensions;
-using Dfe.EarlyYearsQualification.Web.Renderers;
 using Dfe.EarlyYearsQualification.Web.Security;
 using Microsoft.AspNetCore.DataProtection;
 using Microsoft.AspNetCore.Mvc;
@@ -42,7 +42,7 @@ else
     builder.Services.AddTransient<IContentService, ContentfulContentService>();
 }
 
-builder.Services.AddContentfulRenderers();
+builder.Services.AddModelRenderers();
 
 var app = builder.Build();
 app.UseSecureHeadersMiddleware(
